Jarvis_Agent/requirements.txt
google-adk
google-generativeai
fastapi
uvicorn[standard]
python-dotenv
tinytuya
